+++ internal/infra/infra.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/quintans/go-clean-ddd/fake"
 	"github.com/quintans/go-clean-ddd/internal/app/command"
+	"github.com/quintans/go-clean-ddd/internal/app/event"
 	"github.com/quintans/go-clean-ddd/internal/app/query"
 	"github.com/quintans/go-clean-ddd/internal/domain/registration"
 	"github.com/quintans/go-clean-ddd/internal/infra/controller/fakesub"
@@ -14,7 +15,7 @@ import (
 	"github.com/quintans/go-clean-ddd/internal/infra/gateway/fakepub"
 	"github.com/quintans/go-clean-ddd/internal/infra/gateway/postgres"
 	"github.com/quintans/go-clean-ddd/internal/infra/gateway/postgres/ent"
-	"github.com/quintans/go-clean-ddd/lib/event"
+	"github.com/quintans/go-clean-ddd/lib/eventbus"
 	"github.com/quintans/go-clean-ddd/lib/outbox"
 	"github.com/quintans/go-clean-ddd/lib/transaction"
 	"github.com/quintans/toolkit/latch"
@@ -24,7 +25,7 @@ func Start(ctx context.Context, lock *latch.CountDownLatch, cfg Config) {
 	db := NewDB(cfg.DbConfig)
 	defer db.Close()
 
-	bus := event.NewEventBus()
+	bus := eventbus.New()
 	trans := transaction.New[*ent.Tx](
 		bus,
 		func(ctx context.Context) (transaction.Tx, error) {
@@ -45,7 +46,7 @@ func Start(ctx context.Context, lock *latch.CountDownLatch, cfg Config) {
 	createRegistration := command.NewCreateRegistration(registrationWrite, customerRead)
 	confirmRegistration := command.NewConfirmRegistration(registrationWrite)
 
-	bus.AddHandler(registration.EventEmailVerified, command.NewEmailVerifiedHandler(customerWrite, customerRead))
+	bus.AddHandler(registration.EventEmailVerified, event.NewEmailVerifiedHandler(customerWrite, customerRead))
 
 	registrationController := web.NewRegistrationController(createRegistration, confirmRegistration)
 	StartWebServer(ctx, lock, cfg.WebConfig, customerController, registrationController)
@@ -59,7 +60,7 @@ func Start(ctx context.Context, lock *latch.CountDownLatch, cfg Config) {
 	pub := fakepub.NewFakePublisher(mq)
 	outboxMan := outbox.New(trans, 5, pub)
 	outboxMan.Start(ctx, lock, 5*time.Second)
-	bus.AddHandlerF(registration.EventRegistrationCreated, func(ctx context.Context, de event.DomainEvent) error {
+	bus.AddHandlerF(registration.EventRegistrationCreated, func(ctx context.Context, de eventbus.DomainEvent) error {
 		// DEMO: transform the incoming domain event into an integration event if there is a need to.
 		// In this case there is no need
 		return outboxMan.Create(ctx, de)
